Consolidate readiness helpers into rewrite/readiness.rs
Following comment 4 (readiness types moved to their own file), also
move the readiness *helpers* — refresh_candidate_readiness,
resolve_current_readiness, drop_not_ready_after_refresh — from
exploitation.rs into readiness.rs. All three operate exclusively on
CandidateMetadata / RewriteReadiness / ReadinessAnnotatedExec, so
keeping them next to the types they operate on makes readiness cohesive
in one file.

* refresh_candidate_readiness + drop_not_ready_after_refresh are
  pub(super) so exploitation.rs's plan_extension can still call them.
* resolve_current_readiness stays private to readiness.rs (only the
  refresh helper uses it).
* readiness.rs picks up the SessionState / CatalogProviderList /
  cast_to_materialized imports.
* readiness.rs imports RewriteContext from exploitation — one direction
  of the cross-file cycle, which is fine since Rust resolves module
  imports crate-wide.

exploitation.rs drops from 2481 to 2325 lines. readiness.rs grows from
317 to 480 lines and now hosts every readiness-related definition
(4 pub items, 3 helpers, 3 tests).

61/61 lib tests pass, fmt/clippy/doc clean.
